--- constants/localStorage.ts
@@ -1,4 +1,0 @@
-export const localStorageUserTokenKeys = {
-  accessToken: 'gdat',
-  refreshToken: 'gdrt',
-} as const;
